requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-wtf==1.0.0
flask-apscheduler==1.12.3
flask-login==0.5.0
twilio==7.10.0
python-dotenv==0.19.1
werkzeug==2.0.2
pandas==1.3.3
jinja2==3.0.2
python-dateutil==2.8.2
APScheduler==3.8.1
pytest==6.2.5
selenium==4.1.0 